+++ lib/gg_git.dart
@@ -6,15 +6,16 @@
 
 library;
 
-export 'src/gg_git.dart';
+export 'src/base/gg_git_base.dart';
+export 'src/commands/commit_count.dart';
+export 'src/commands/commit.dart';
 export 'src/commands/get_tags.dart';
-export 'src/commands/is_committed.dart';
-export 'src/commands/is_pushed.dart';
+export 'src/commands/has_remote.dart';
 export 'src/commands/head/head_hash.dart';
 export 'src/commands/head/head_message.dart';
-export 'src/commands/modified_files.dart';
-export 'src/commands/commit.dart';
-export 'src/commands/commit_count.dart';
+export 'src/commands/is_committed.dart';
+export 'src/commands/is_pushed.dart';
 export 'src/commands/last_changes_hash.dart';
-export 'src/commands/has_remote.dart';
-export 'src/base/gg_git_base.dart';
+export 'src/commands/modified_files.dart';
+export 'src/commands/upstream_branch.dart';
+export 'src/gg_git.dart';